requirements.txt
classification-models-3D==1.0.2
dash-bootstrap-components==0.13.1
dash==2.0.0
Flask==2.0.2
ipywidgets==7.6.5
Keras-Applications==1.0.8
lime==0.2.0.1
matplotlib==3.4.3
nibabel==3.2.1
numpy==1.19.5
opencv-python==4.5.3.56
pandas==1.3.3
Pillow==8.4.0
plotly==5.3.1
scikit-image==0.18.3
scipy==1.7.1
shapely==1.7.1
tensorflow==2.6.0
tqdm==4.62.3
fpdf==1.7.2